ISSUE #3882 - rename components
