--- frontend/src/v5/ui/controls/modal/formModal/formModalHeader/formModalHeader.component.tsx
@@ -1,40 +1,0 @@
-/**
- *  Copyright (C) 2023 3D Repo Ltd
- *
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU Affero General Public License as
- *  published by the Free Software Foundation, either version 3 of the
- *  License, or (at your option) any later version.
- *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU Affero General Public License for more details.
- *
- *  You should have received a copy of the GNU Affero General Public License
- *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-import CloseIcon from '@assets/icons/outlined/close-outlined.svg';
-import { CloseButton, Header, Subtitle, Title } from './formModalHeader.styles';
-
-type IFormModalHeader = {
-	title: string;
-	subtitle?: string;
-	handleClose: () => void;
-	disableClosing?: boolean;
-};
-
-export const FormModalHeader = ({ title, subtitle, handleClose, disableClosing }: IFormModalHeader) => (
-	<Header>
-		<div>
-			<Title>
-				{title}
-			</Title>
-			{subtitle && <Subtitle>{subtitle}</Subtitle>}
-		</div>
-		<CloseButton aria-label="Close dialog" onClick={handleClose} disabled={disableClosing}>
-			<CloseIcon />
-		</CloseButton>
-	</Header>
-);
