+++ frontend/src/v5/ui/routes/dashboard/projects/containers/uploadFileForm/uploadFileForm.styles.ts
@@ -18,7 +18,7 @@
 import styled from 'styled-components';
 import { DragAndDrop } from '@controls/dragAndDrop';
 import { DashboardListHeader } from '@components/dashboard/dashboardList/dashboardListHeader';
-import { FormModal } from '@controls/headedModal/formModal/formModal.component';
+import { FormModal } from '@controls/formModal/formModal.component';
 import { ScrollArea } from '@controls/scrollArea';
 
 const MODAL_PADDING = 35;
